Refactored AuthProcedureTest to separate primary unit tests and scenarios.
